Changed binary format for 3.1.05.
* String encoding simplified for reading by using number of bytes rather than number of UTF-8 characters.
* Parent bone index is no longer minus 1 and omitted for root.
* Mesh and weighted mesh vertices array is no longer preceeded by a length. UVs array length is vertex count, not vertex array length. Edges changed to short.
* Bounding box vertices array length is vertex count, not vertex array length.

+++ spine-libgdx/spine-libgdx/src/com/esotericsoftware/spine/SkeletonJson.java
@@ -300,7 +300,7 @@ public class SkeletonJson {
 				mesh.updateUVs();
 
 				if (map.has("hull")) mesh.setHullLength(map.require("hull").asInt() * 2);
-				if (map.has("edges")) mesh.setEdges(map.require("edges").asIntArray());
+				if (map.has("edges")) mesh.setEdges(map.require("edges").asShortArray());
 			} else {
 				mesh.setInheritFFD(map.getBoolean("ffd", true));
 				linkedMeshes.add(new LinkedMesh(mesh, map.getString("skin", null), slotIndex, parent));
@@ -343,7 +343,7 @@ public class SkeletonJson {
 				mesh.updateUVs();
 
 				if (map.has("hull")) mesh.setHullLength(map.require("hull").asInt() * 2);
-				if (map.has("edges")) mesh.setEdges(map.require("edges").asIntArray());
+				if (map.has("edges")) mesh.setEdges(map.require("edges").asShortArray());
 			} else {
 				mesh.setInheritFFD(map.getBoolean("ffd", true));
 				linkedMeshes.add(new LinkedMesh(mesh, map.getString("skin", null), slotIndex, parent));
