+++ spine-libgdx/spine-libgdx/src/com/esotericsoftware/spine/attachments/MeshAttachment.java
@@ -53,7 +53,7 @@ public class MeshAttachment extends Attachment implements FfdAttachment {
 	private boolean inheritFFD;
 
 	// Nonessential.
-	private int[] edges;
+	private short[] edges;
 	private float width, height;
 
 	public MeshAttachment (String name) {
@@ -182,11 +182,11 @@ public class MeshAttachment extends Attachment implements FfdAttachment {
 		this.hullLength = hullLength;
 	}
 
-	public int[] getEdges () {
+	public short[] getEdges () {
 		return edges;
 	}
 
-	public void setEdges (int[] edges) {
+	public void setEdges (short[] edges) {
 		this.edges = edges;
 	}
